+++ src/renderer/controllers/jwt-controller.js
@@ -20,7 +20,8 @@ module.exports = class JwtController {
 
   saveJwt(jwt = {
     address: '',
-    accessToken: ''
+    accessToken: '',
+    rewardsEligibility: {}
   }) {
     this.state.saved.auth = jwt;
     dispatch('stateSaveImmediate');
@@ -39,22 +40,22 @@ module.exports = class JwtController {
     }
   }
 
-  // Shows a modal saying that we have an update
   validateJwt(jwt) {
     if (!jwt) {
       console.log('No JWT found');
       return;
     }
     JwtApi.validateJwt(jwt)
-      .then(({success, res, address }) => {
+      .then(({success, res, address, rewardsEligibility }) => {
         if (success) {
           return dispatch('saveJwt', {
             address: address,
-            accessToken: jwt
+            accessToken: jwt,
+            rewardsEligibility
           });
         }
         if (res.statusCode === 401) dispatch('saveJwt');
-        dispatch('error', new Error(`Error restoring session, status code: ${res.statusCode} ${res.message}`));
+        else dispatch('error', new Error(`Error restoring session, status code: ${res.statusCode} ${res.message}`));
       })
       .catch(err => {
         dispatch('error', err);