+++ src/renderer/pages/torrent-list-page.js
@@ -32,7 +32,7 @@ module.exports = class TorrentList extends React.Component {
     contents.push(...torrentElems)
     contents.push(
       <div key='torrent-placeholder' className='torrent-placeholder'>
-        <span className='ellipsis'>Drop a torrent file here or paste a magnet link</span>
+        <span className='ellipsis'>Paste a magnet link</span>
       </div>
     )